requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.59.2
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.1
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
reportlab==4.4.3
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
gunicorn==21.2.0
whitenoise==6.7.0

